feat: add site generator class
