+++ bin/writr
@@ -4,14 +4,6 @@
 
 const Writr = require("../dist/index").Writr;
 
-const writr = new Writr;
+const writr = new Writr();
 
 writr.parseCLI(process);
-
-if(writr.command === 'build') {
-  writr.runCLI().then(() => {
-    process.exit();
-    console.info("Done!");
-  })
-}
-
