+++ src/index.ts
@@ -1,25 +1,19 @@
-import * as del from "del";
-import * as fs from "fs-extra";
 import { createCommand } from "commander";
 
 import {DataService} from "./data/dataService";
 import {Config} from "./config";
-import {HtmlRenderProvider} from "./render/htmRenderlProvider";
-import {JSONRenderProvider} from "./render/jsonRenderProvider";
-import {AtomRenderProvider} from "./render/atomRenderProvider";
-import {ImageRenderProvider} from "./render/imageRenderProvider";
-import {Migrate} from "./migrate";
 import {Setup} from "./utils/setup";
 import {ConsoleMessage} from "./log";
 import {Serve} from "./serve";
+import {SiteGenerator} from "./generator";
 
 export class Writr {
 
   config: Config | undefined;
   data: DataService | undefined;
-  command: string | undefined;
+  result = false;
 
-  parseCLI(process: NodeJS.Process) {
+  async parseCLI(process: NodeJS.Process) {
 
     const program = createCommand();
 
@@ -33,28 +27,13 @@ export class Writr {
       .option("-r, --render <list>", "What do you want rendered such as html or json (example --render html,json)")
       .option("-c, --config <path>", "custom configuration path")
       .option("-m, --migrate <type> <source> <destination>", "Migrate from Jekyll to Writr")
-      .action((options: any) => {
-        this.command = "build";
-
-        const params = options.opts();
-
-        this.config = new Config();
-
-        if (params.config) {
-          this.config.loadConfig(params.config);
+      .action(async (options: any) => {
+        try{
+          await new SiteGenerator(options).run();
+          this.result = true;
+        } catch (error: any) {
+          new ConsoleMessage().error('Error: '+ error.message);
         }
-
-        if (params.path) {
-          this.config.loadPath(params.path);
-        }
-
-        if (params) {
-          this.config.loadParams(params)
-        }
-
-        this.config.loadProgram(options);
-
-        this.data = new DataService(this.config);
       })
 
     program
@@ -63,8 +42,8 @@ export class Writr {
       .argument('[name]', 'Name of the project', 'Blog')
       .action(async (name: string) => {
         try{
-          this.command = "init";
           await new Setup(name).init();
+          this.result = true;
         } catch (error: any) {
           new ConsoleMessage().error('Error: '+ error.message);
         }
@@ -75,8 +54,8 @@ export class Writr {
       .description('Create new markdown file')
       .action(async() => {
         try{
-          this.command = "new";
           await new Setup('new').new();
+          this.result = true;
         } catch (error: any) {
           new ConsoleMessage().error('Error: '+ error.message);
         }
@@ -91,9 +70,9 @@ export class Writr {
       .option("-w, --watch", "Watch for changes and rebuild", false)
       .action(async(options: any) => {
         try{
-          this.command = "serve";
           const params = options.opts();
           await new Serve(params).run();
+          this.result = true;
         } catch (error: any) {
           new ConsoleMessage().error('Error: '+ error.message);
         }
@@ -101,49 +80,6 @@ export class Writr {
 
     program.parse(process.argv);
 
-  }
-
-  async runCLI(): Promise<boolean> {
-    let result = true;
-
-    if (this.data === undefined || this.config === undefined) {
-      return false;
-    }
-
-    const {migrate} = this.config.params;
-
-    if (migrate) {
-      const [src, dest] = this.config.program.args;
-      await new Migrate(migrate).migrate(src, dest);
-      return true;
-    }
-
-    if (fs.existsSync(this.config.output)) {
-      del.sync(this.config.output);
-    }
-
-    let render: boolean | undefined = true;
-
-    for (let i = 0; i < this.config.render.length; i++) {
-      let type = this.config.render[i];
-      if (type === "html") {
-        render = await new HtmlRenderProvider().render(this.data, this.config);
-      }
-      if (type === "json") {
-        render = await new JSONRenderProvider().render(this.data, this.config);
-      }
-      if (type === "atom") {
-        render = await new AtomRenderProvider().render(this.data, this.config);
-      }
-      if (type === "images") {
-        render = await new ImageRenderProvider().render(this.data, this.config);
-      }
-    }
-
-    if (render) {
-      result = render;
-    }
-
-    return result;
+    return this.result;
   }
 }
